// ui/src/lib/pocketbase-types.ts
export type IsoDateString = string
export type RecordIdString = string
export type HTMLString = string
// ...
export type HopsRecord<Theaders = unknown, Tipinfo = unknown> = {
	body?: HTMLString
	created?: IsoDateString
	headers?: null | Theaders
	hop_number: number
	host?: string
	http_version?: string
	id: string
	ipaddr?: string
	ipinfo?: null | Tipinfo
	path?: string
	scheme?: string
	search_id?: RecordIdString
	status_code: number
	status_phrase?: string
	time_elapsed?: number
	updated?: IsoDateString
	url: string
}